+++ src/Cwaff-Misc/CwaffPatches.cs
@@ -454,3 +454,31 @@ static class AllowMoreThan30KnockbackPatch
     return KnockbackUnleasher.UNLEASHED_KNOCKBACK_CAP;
   }
 }
+
+/// <summary>Allow beams to apply arbitrary status effects (used by Macchi Auto and Photon Cannon)</summary>
+[HarmonyPatch(typeof(BasicBeamController), nameof(BasicBeamController.FrameUpdate))]
+static class BeamApplyArbitraryStatusEffectPatch
+{
+    [HarmonyILManipulator]
+    private static void BeamApplyArbitraryStatusEffectIL(ILContext il)
+    {
+        ILCursor cursor = new ILCursor(il);
+
+        if (!cursor.TryGotoNext(MoveType.Before,
+          instr => instr.MatchCall<BraveBehaviour>("get_projectile"),
+          instr => instr.MatchLdfld<Projectile>("AppliesSpeedModifier")))
+            return;
+
+        // BeamController is already on the stack here
+        cursor.Emit(OpCodes.Ldloc_S, (byte)30); // V_30 == the enemy gameActor
+        cursor.CallPrivate(typeof(BeamApplyArbitraryStatusEffectPatch), nameof(ApplyExtraBeamStatusEffects));
+        cursor.Emit(OpCodes.Ldarg_0); // put the BeamController back on the call stack
+    }
+
+    private static void ApplyExtraBeamStatusEffects(BeamController beam, GameActor gameActor)
+    {
+        foreach (GameActorEffect e in beam.projectile.statusEffectsToApply)
+            if (UnityEngine.Random.value < BraveMathCollege.SliceProbability(beam.statusEffectChance, BraveTime.DeltaTime))
+                gameActor.ApplyEffect(e);
+    }
+}